Fix deadlock from Stop() called from callbacks
Calling Stop() from callbacks may result in a deadlock if the
callback is called because a message is received. Stop() will
wait for receiver to stop, but receiver can't stop until the
callback is returned.

This change allow Stop() to initiate stopping and return without
waiting for Stop() to finish.

This ability is important for implementing use cases where
connection settings offer is received and the callback needs
to Stop() the current OpAMP client and reconnect again using
the new settings.